pyproject.toml
[tool.poetry]
name = "ai-mafia"
version = "0.0.1-alpha"
description = ""
authors = ["voorhs <ilya_alekseev_2016@list.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
chatsky = {extras = ["telegram"], version = "^0.9.0"}
python-dotenv = "^1.0.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.3"
pytest = "^8.3.4"

[tool.ruff]
line-length = 120
indent-width = 4
exclude = ["experiments", "scripts"]
target-version = "py310"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "TD",      # todos
    "FIX",     # fixmes
    "S311",    # random usage
    "COM812",  # trailing commas
    "ISC001",  # implicit string concatenation
    "ERA001",  # commented code
    "FBT",     # boolean trap
    "D",       # docstrings
    "ANN",     # type annotations
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "D104"]
"ai_mafia/__init__.py" = ["I001"]
"tests/*.py" = ["S", "PLR2004", "ERA", "D", "ANN", "SLF"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra" # `--cov` option breaks pycharm's test debugger
testpaths = [
    "tests",
]
pythonpath = "ai_mafia"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
